♻️ Integrate genStoryFile and runGenStoryFile
