--- packages/auto-story-generator/.release-it.cjs
@@ -1,18 +1,0 @@
-module.exports = {
-  npm: {
-    publish: true,
-  },
-  /* github: {
-    release: true,
-  }, */
-  git: {
-    tag: false,
-    requireCleanWorkingDir: false,
-    commitMessage:
-      ":bookmark: release @takuma-ru/auto-story-generator@${version}",
-  },
-  /* hooks: {
-    "after:git:release": "git pull origin ${version}",
-  }, */
-  plugins: {},
-};
