+++ packages/auto-story-generator/src/types/ErrorType.ts
